api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
groq==0.8.0
sentence-transformers==2.7.0
chromadb==0.5.23
sqlalchemy==2.0.30
httpx==0.27.0